Word Card List Shuffle Работает экзамен

+++ app/src/main/java/com/greenkeycompany/exam/fragment/TrainingCompletedUtil.java
@@ -6,8 +6,8 @@ package com.greenkeycompany.exam.fragment;
 
 public class TrainingCompletedUtil {
 
-    private static final float COMPLETED_LIMIT_PERCENT = 70;
-    private static final float MAX_PERCENT = 100;
+    private static final int COMPLETED_LIMIT_PERCENT = 70;
+    private static final int MAX_PERCENT = 100;
 
     public static boolean isCompleted(int trueAnswerCount, int wordCardCount) {
         if (trueAnswerCount * MAX_PERCENT >= COMPLETED_LIMIT_PERCENT * wordCardCount)
